Zip it up (#2)
* Add vendor_kernel_boot files to target-files-package

* This is a required dependency of target-files-package
so that it can be built from a clean build.

Test: rm -rf out/target/product && m target-files-package
Change-Id: I89e60f47e3e2cd6029a92b1aca505acd9cf4645b

* disable ART JIT and JIT profiling

Signed-off-by: anupritaisno1 <www.anuprita804@gmail.com>

* disable BOARD_USES_SYSTEM_OTHER_ODEX for mainline

The system_other image is gone after the first update and the OS ends up
needing to compile the code for these apps as part of performing each OS
update. GrapheneOS doesn't use JIT compilation and the AOT compilation
isn't guided by JIT profiles so this ends up using the same amount of
space and doing a bunch of unnecessary work on each device. It also
reduces the scope of verified boot by having the compiled odex code for
base system apps outside of the verified system partition.

Signed-off-by: anupritaisno1 <www.anuprita804@gmail.com>

* Revert "Add a build-time check for dexpreopting system server jars."

This reverts commit b6931f2f5bd80249c02f466a4ea45fc750c7e761.

Signed-off-by: anupritaisno1 <www.anuprita804@gmail.com>
Co-authored-by: Chirayu Desai <chirayudesai1@gmail.com>
Co-authored-by: Daniel Micay <danielmicay@gmail.com>
Co-authored-by: hurtcopain <jacksummers715@gmail.com>